pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aurix"
version = "0.1.0"
description = "Enterprise platform for confident automation of agentic workflows"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Aurix Team" },
]
keywords = [
    "automation",
    "ai",
    "agents",
    "workflow",
    "confidence",
    "risk-assessment",
    "devops",
    "sdlc",
    "code-review",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "pyyaml>=6.0.0",
    "openai>=1.10.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]

[project.urls]
Documentation = "https://github.com/palashroy/aurix#readme"
Issues = "https://github.com/palashroy/aurix/issues"
Source = "https://github.com/palashroy/aurix"

[project.scripts]
aurix = "aurix.cli.commands:app"

[project.entry-points."aurix.modules"]
code_review = "aurix.modules.code_review:CodeReviewModule"
sdlc = "aurix.modules.sdlc:SDLCModule"

[tool.hatch.build.targets.wheel]
packages = ["aurix"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source_pkgs = ["aurix"]
branch = true
parallel = true
omit = [
    "aurix/__about__.py",
]

[tool.coverage.paths]
aurix = ["aurix"]
tests = ["tests"]

[tool.coverage.report]
exclude_lines = [
    "no cov",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "pass",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # Pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
    "ARG",   # flake8-unused-arguments
    "SIM",   # flake8-simplify
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise without from inside except
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"tests/*" = ["ARG001"]

[tool.ruff.isort]
known-first-party = ["aurix"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.black]
target-version = ["py311"]
line-length = 100
